firmware: app: tasks: Refactor operation control task to use direct task notifications
The rationale for the change is based on the excessive amount of
locks/unlocks required to manage the satellite state using only the
sat_data_buf instance. Also, the change reduces wasting computing
resources, because tasks now block on xTaskNotifyWait calls instead of
wasting time on if statements, checking if they are allowed to run.