firmware: app: tasks: housekeeping: Fix obdh data save functionality
Because the system time is controlled outside the housekeeping task, the
only way that the call (system_get_time() % 60U) returns 0 is if the
housekeeping task executes exactly after the time_control task when it
update the system time to a "multiple of a minute". Therefore, the
housekeeping task definitely did not save the data to fram every minute.
Now, with the changes, it does actually saves it in the expected time.